Add log file
This patch adds a log file for the app in /tmp.  Once the application is
closed, the log file is deleted.  If an unexpected error occurs, the
user has the option to save the log file so that it can be submitted to
us.  The same option is available in the about dialog.

Fixes: https://github.com/Nitrokey/nitrokey-app2/issues/74