add WelcomeTab 	modified:   nitrokeyapp/gui.py 	new file:   nitrokeyapp/ui/icons/s_icon_Logo_App.png 	modified:   nitrokeyapp/ui/mainwindow.ui 	new file:   nitrokeyapp/ui/welcome_tab.ui 	new file:   nitrokeyapp/welcome_tab.py
A WelcomeTab that is displayed when the app is started and also when the key is removed. It contains a small explanation of what to do and also the current app version. Furthermore the app logo and a link to the documentation.
	modified:   nitrokeyapp/gui.py
	new file:   nitrokeyapp/ui/icons/s_icon_Logo_App.png
	modified:   nitrokeyapp/ui/mainwindow.ui
	new file:   nitrokeyapp/ui/welcome_tab.ui
	new file:   nitrokeyapp/welcome_tab.py

code style change
	modified:   nitrokeyapp/welcome_tab.py

code style change
	modified:   nitrokeyapp/welcome_tab.py

code style change
	modified:   nitrokeyapp/welcome_tab.py

code style change
	modified:   nitrokeyapp/gui.py
	modified:   nitrokeyapp/welcome_tab.py

code style change
	modified:   nitrokeyapp/gui.py

code style change
	modified:   nitrokeyapp/gui.py