Bump pynitrokey, mypy
This patch updates pynitrokey to 0.4.40 and mypy to 1.4.

pynitrokey 0.4.39 updated the sdsk dependency which fixes a build issue
on some systems.  0.4.40 also fixed the “generator did not stop after
throw” issue that has also been reported for the app.

mypy 1.4 fixes an occasional type checking error that had to be fixed by
removing the cache.

Fixes: https://github.com/Nitrokey/nitrokey-app2/issues/144