Use poetry run and poetry env remove in Makefile
Previously, we determined the venv path used by poetry and used it to
call the executables directly.  This has adds an additional poetry
invocation for all targets even if poetry is not used at all.  It even
slows down other uses of the Makefile, e. g. for auto-completion.
Instead, we can use poetry run and poetry env remove to avoid this
overhead.